Integrate with Approval Service
Uses the Approval API client to send approval request to
the approval service and store the request reference for each order
item. When the approval request is created it includes the name of
the portfolio item and also the parameters to be used in provisioning.
Any password parameters are masked