Added support for Provider Control Parameters
For each provider (source) we need specific information related to
provisioning. For e.g. on Openshift platform we need to let the user
pick a project (namespace) where the provision should happen.
This is a temporary solution till we can get the topology service to
provide us the provider control parameters.

We pass in a portfolio_item_id and it returns a JSON schema that the
UI can render