When updating invalid parameters return 400
The earlier spec was expecting that invalid parameters should
return a 200 and only update the attributes that are needed.
Extra attributes are ignored. Now that we have openapi parser
every parameter needs to be defined in the openapi.json.
Extra attributes lead to a 400 error.
Removed the specs which are no longer relevant.