Removed the exceptions app
This app was added in PR https://github.com/ManageIQ/catalog-api/pull/301
because we had lot of rescue froms in all of our controllers. This
deviates from the pattern established in topology and sources api. We
now have a better way of error reporting that uses the "cause" in an
exception object to build a chain of error messages. It also collapses
all of the rescue_from into one for StandardError and RuntimeError and
since most of the errors inherit from one of these we can capture them
in one place in the common gem.
The PR that did this was https://github.com/ManageIQ/manageiq-api-common/pull/114