Seed for org admin per tenant
(1) Create an ITSM Portfolio
(2) Create/Update a group called **Default Catalog & Approval Administrator**
(3) If the following 2 roles are missing add them to the group
    * **Catalog Administrator**
    * **Approval Administrator**
(4) Add the current user to the Group
(5) If the seeding was done send back HTTP Status Code 201 (Created) and
create a record in the RbacSeed table for the tenant.
(6) If the seeding was skipped send back HTTP Status Code 204 (No
content)

The seeding is a 3 step process for the UI
 * GET https://<CATALOG_URL>/api/catalog/v1.3/tenants
 * Parse the Tenant ID from the response
 * POST https://<CATALOG_URL>/api/catalog/v1.3/tenants/<<TENANT_ID>>/seed

If the seeding was done you will get back a 201, the caller should
reload RBAC permissions

If the seeding was skipped you will get back a 204, no need to reload
RBAC permissions

If the user is not an org admin you will get back a 403.